tries to add is_habitable_atmos proc
Checks if the turf has sufficent atmos for someones habitability, you set the requirements
Returns TRUE if atmos at the turf has more then min or/and less then max of gasses.
Returns FALSE if atmos at the turf has less the min or/and more then max of gasses.